requirements.txt
opencv-python
python-dotenv
pywin32
twilio
cloudinary